project-wrapper-api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
click==7.1.2              # via flask
colour-sort==0.1.3        # via wrapper-api (setup.py)
flask==1.1.2              # via wrapper-api (setup.py)
gunicorn==20.0.4          # via wrapper-api (setup.py)
importlib-resources==1.5.0  # via colour-sort
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
markupsafe==1.1.1         # via jinja2
numpy==1.19.2             # via colour-sort
pillow==6.2.2             # via colour-sort
werkzeug==1.0.1           # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
